refactor(github): extract MCP result envelope from helpers
Move normalize_github_tool_result and its private JSON fallback into
integrations/github/envelope.py, matching groundcover's envelope role.
Leave availability, credentials, and MCP config resolution in helpers.py.

Fixes #5631
